Station Filter Selection Screen Implementation
- Distance slider is implemented
- Filter implemented
- App name is changed

+++ charger_app/charger_app/Modules/StationFilter/FilterSettings/FilterSettings.swift
@@ -10,6 +10,6 @@ import Foundation
 class FilterSettings {
     var chargeType: [ChargeType]? = [ChargeType]()
     var socketTpye: [SocketType]? = [SocketType]()
-    var distance: Double?
+    var distance: Int?
     var serviceType: [ServiceType]? = [ServiceType]()
 }
